更新gulp套件到4.0 修正MSDeploy如果沒發生錯誤會找不到error log檔案

+++ msdeploy/MSDeployExecute.ps1
@@ -3,28 +3,30 @@ param()
 
 Trace-VstsEnteringInvocation $MyInvocation
 try {
-    Import-VstsLocStrings "$PSScriptRoot\task.json"
+	Import-VstsLocStrings "$PSScriptRoot\task.json"
 
-    [string]$verb = Get-VstsInput -Name verb
+	[string]$verb = Get-VstsInput -Name verb
 	[string[]]$otherArgs = (Get-VstsInput -Name otherArgs).Split("`n");
 
-    [System.Console]::OutputEncoding = [System.Text.Encoding]::UTF8
+	[System.Console]::OutputEncoding = [System.Text.Encoding]::UTF8
 
-    $tempFile = [guid]::NewGuid()
-    $hasError = $false
+	$tempFile = [guid]::NewGuid()
+	$hasError = $false
 
-    & "$PSScriptRoot\MSDeploy.cmd" -verb:$verb $otherArgs "2>>$tempFile.error" | Out-Host
+	& "$PSScriptRoot\MSDeploy.cmd" -verb:$verb $otherArgs "2>>$tempFile.error" | Out-Host
 
-    gc "$tempFile.error" -Encoding UTF8 | %{
-        $hasError = $true
-        Write-Warning $_
-    }
-    Remove-Item "$tempFile.error"
+	if (Test-Path "$tempFile.error" -PathType Leaf) {
+		gc "$tempFile.error" -Encoding UTF8 | % {
+			$hasError = $true
+			Write-Warning $_
+		}
+		Remove-Item "$tempFile.error"
+	}
 
-    if($hasError -eq $true)
-    {
-        Write-Error "MsDeploy發生錯誤"
-    }
-} finally {
-    Trace-VstsLeavingInvocation $MyInvocation
+	if ($hasError -eq $true) {
+		Write-Error "MsDeploy發生錯誤"
+	}
+}
+finally {
+	Trace-VstsLeavingInvocation $MyInvocation
 }
